Added the Kage2D engine.
Added a grid header file and class file, that are used to define all the variables and then create the functions required to draw the main grid.

Defined some named variables in the example header file, that will be linked to the tile textures saved in the bin folder.

Added image buttons that should theoretically allow the user to select any tile from the available selection and place them on the grid.

Known Errors:

1. The tiles are not being placed on the grid itself.
2. The tiles are drawn in the top-left corner of the window and are not the same size as an individual cell from the grid.
3. Changing tiles is a bit... finicky. Sometimes changing colors and placing tiles will be shown by a change in the top-left tile, sometimes it doesnt. Sometimes selecting blue results in the tile changing color instantly before I can even click on the grid itself.