Renamed the Sprites variable to tileTexture in order to load my tile image files into it.
Created a LoadMap function which stores the for loop required to place the tiles in the grid as well as changing the tileId from 0 (meaning empty tile) to match that of the tile being placed. The for loop for this function was moved from the example.cpp file into the tile.cpp file.

Added the Load and LoadMap functions in the Start function in the example.cpp script.