Added comments. Lotta comments.
Added a tile.h and tile.cpp script.
Moved the tilemap array to the header file.
Fixed a previous issue where the tiles weren't being placed on the grid.

Modified the Macros in the grid.h file, so that the grid size can be modified by simply specifying the no. of cells in a row/column.

Added more colored tiles to choose from.
Created a separate UI menu that lets the user save the grid in a text file after working on it.
It also lets the user load the file, but atm it cannot load the tiles onto the grid itself. It only displays the tileIDs in their respective positions in the console window.
There is also a text bar that lets the user name the save file, but at the moment, it does not work with the load function. [Currently Commented out]

KNOWN ISSUES:

Cannot load tiles onto the grid.
Cannot rename the save file from within the editor.
Did not add a custom background.